refactor(stream-proxy): split stream module
